requirements.txt
altgraph==0.17
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.1
attrs==19.3.0
audioread==2.1.9
Automat==20.2.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.0
cssselect==1.1.0
cycler==0.10.0
decorator==5.0.5
distlib==0.3.2
Django==2.2
django-admin-view-permission==1.9
django-cors-headers==3.7.0
django-haystack==2.8.1
django-redis==4.12.1
djangorestframework==3.12.2
dnspython==1.16.0
easygui==0.98.2
et-xmlfile==1.0.1
eventlet==0.31.0
fakeredis==1.5.2
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.0.12
future==0.18.2
greenlet==1.1.0
importlib-resources==5.2.0
itchat==1.3.10
jdcal==1.4.1
jieba==0.42.1
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
kombu==5.0.2
librosa==0.8.0
llvmlite==0.36.0
lupa==1.9
lxml==4.5.2
matplotlib==3.3.4
MouseInfo==0.1.3

parsel==1.6.0
pefile==2019.4.18
Pillow==7.2.0
pooch==1.3.0
prompt-toolkit==3.0.17
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
pycparser==2.20
PyDispatcher==2.0.5
PyEmail==0.0.1
pygame==2.0.0
PyGetWindow==0.0.8
PyHamcrest==2.0.2
pyinstall==0.1.4
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
PyMsgBox==1.0.8
PyMySQL==0.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
pypng==0.0.20
PyQRCode==1.2.1
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.5
pywin32==228
pywin32-ctypes==0.2.0
queuelib==1.5.0
redis==3.2.0
requests==2.24.0
resampy==0.2.2
scikit-learn==0.24.1
scipy==1.5.4
Scrapy==2.3.0
selenium==3.141.0
service-identity==18.1.0
setuptools-scm==6.0.1
six==1.15.0
sortedcontainers==2.4.0
sounddevice==0.4.1
SoundFile==0.10.3.post1
sqlparse==0.4.1
threadpoolctl==2.1.0
Twisted==20.3.0
typing-extensions==3.7.4.3
UNKNOWN==0.0.0
urllib3==1.25.10
vine==5.0.0
virtualenv==20.4.7
w3lib==1.22.0
wcwidth==0.2.5
Whoosh==2.7.4
xlrd==1.2.0
zipp==3.4.1
zmail==0.2.8
zope.interface==5.1.0